pyproject.toml
[project]
name = "model_manager"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Arthur Chi", email = "arthur@openfoundry.ai"},
]
dependencies = [
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
    "sagemaker>=2.218.1",
    "boto3>=1.34.99",
    "setuptools>=69.5.1",
    "huggingface-hub>=0.23.0",
    "pydantic>=2.7.1",
    "datasets>=2.19.1",
    "transformers>=4.40.2",
    "inquirer>=3.2.4",
    "inquirerpy>=0.3.4",
    "uvicorn>=0.29.0",
    "fastapi>=0.111.0",
    "litellm>=1.37.5",
    "cryptography>=42.0.7",
    "supabase>=2.4.5",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
test = [
    "openai>=1.27.0",
    "pytest>=8.2.0",
    "httpx>=0.27.0",
]
